refactor: don't depend on @Inject for _MatMenuBase DI (#15975)
In Ivy, undecorated base classes (like _MatMenuBase) don't have the same
behavior as View Engine; specifically, inheriting the constructor from the
base class will not work in the DI system. Even in View Engine this was
often counter-intuitive - only constructors with DI decorators would be
properly inherited in JIT.

This commit updates _MatMenuBase to not rely on this implementation detail
of Angular, and instead moves the injected constructor onto _MatMenu itself.