fix(propover-edit): incorrect template root note retrieved in ivy
Currently the popover-edit hover directive retrieves the root node
of the embedded view for the popover by looking at the next sibling
of the directive host element. This worked fine in view engine where
the inserted view was added after the current host element, but in
Ivy, the embedded view seems to be inserted before the directive
host node (which is a comment node).

The popover directive should not rely on the implementation detail
of the embedded view insertion and rather retrieve the embedded
view root node through the `EmbeddedViewRef`.