build: fix tslint being unable to type check cross-package references
Since we use a `tsconfig.json` for TSLint which does not have any path
mappings, TSlint is unable to use the `ts.TypeChecker` for nodes which
are declared in a separate entry-point or package. This causes custom
TSLint rules such as the `undecorated-base-class-di`  rule to accidentally
ignore issues such as that `MatCheckbox` (from mdc-checkbox) depends on
the undecorated base class `MatMenu`.